+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Adding Simpl display device
+ - Adding Simpl display with audio device
 
 ### Removed
  - Removed old, unused SPlus display interface project